Latest filesize improvements
