+++ dev/react/src/tests/animate-style-autoplay.tsx
@@ -13,7 +13,7 @@ export const App = () => {
             { duration: 0.1, autoplay: false }
         )
 
-        return () => animation.stop()
+        return () => animation.cancel()
     }, [])
 
     return <div id="box" ref={ref} style={style} />
