+++ dev/react/src/tests/animate-style-duration.tsx
@@ -33,7 +33,7 @@ export const App = () => {
         }
 
         return () => {
-            animation.stop()
+            animation.cancel()
         }
     }, [])
