+++ dev/react/src/tests/animate-style-play.tsx
@@ -16,7 +16,7 @@ export const App = () => {
         animation.pause()
         animation.play()
 
-        return () => animation.stop()
+        return () => animation.cancel()
     }, [])
 
     return <div id="box" ref={ref} style={style} />
