+++ dev/react/src/tests/animate-style-promise.tsx
@@ -18,7 +18,7 @@ export const App = () => {
             ref.current.style.backgroundColor = "red"
         })
 
-        return () => animation.stop()
+        return () => animation.cancel()
     }, [])
 
     return <div id="box" ref={ref} style={style} />
