+++ dev/react/src/tests/animate-style.tsx
@@ -13,7 +13,7 @@ export const App = () => {
             { duration: 0.1 }
         )
 
-        return () => animation.stop()
+        return () => animation.cancel()
     }, [])
 
     return <div id="box" ref={ref} style={style} />
