+++ packages/framer-motion/src/animation/interfaces/visual-element-target.ts
@@ -6,7 +6,7 @@ import type { VisualElementAnimationOptions } from "./types"
 import { animateMotionValue } from "./motion-value"
 import { setTarget } from "../../render/utils/setters"
 import { AnimationPlaybackControls } from "../types"
-import { getValueTransition } from "../utils/transitions"
+import { getValueTransition } from "../utils/get-value-transition"
 import { frame } from "../../frameloop"
 import { getOptimisedAppearId } from "../optimized-appear/get-appear-id"
 import { addValueToWillChange } from "../../value/use-will-change/add-will-change"
