+++ packages/framer-motion/src/animation/utils/__tests__/transitions.test.ts
@@ -1,4 +1,4 @@
-import { isTransitionDefined } from "../transitions"
+import { isTransitionDefined } from "../is-transition-defined"
 
 describe("isTransitionDefined", () => {
     test("Detects a transition", () => {
