+++ packages/framer-motion/src/animation/utils/get-value-transition.ts
@@ -1,0 +1,9 @@
+import { Transition } from "../../types"
+
+export function getValueTransition(transition: Transition, key: string) {
+    return (
+        transition[key as keyof typeof transition] ||
+        (transition as any)["default"] ||
+        transition
+    )
+}
