+++ packages/framer-motion/src/render/dom/value-types/number.ts
@@ -1,79 +1,13 @@
-import { alpha, scale } from "../../../value/types/numbers"
-import {
-    degrees,
-    progressPercentage,
-    px,
-} from "../../../value/types/numbers/units"
+import { alpha } from "../../../value/types/numbers"
+import { browserNumberValueTypes } from "./number-browser"
+import { transformValueTypes } from "./transform"
 import { int } from "./type-int"
 import { ValueTypeMap } from "./types"
 
 export const numberValueTypes: ValueTypeMap = {
-    // Border props
-    borderWidth: px,
-    borderTopWidth: px,
-    borderRightWidth: px,
-    borderBottomWidth: px,
-    borderLeftWidth: px,
-    borderRadius: px,
-    radius: px,
-    borderTopLeftRadius: px,
-    borderTopRightRadius: px,
-    borderBottomRightRadius: px,
-    borderBottomLeftRadius: px,
-
-    // Positioning props
-    width: px,
-    maxWidth: px,
-    height: px,
-    maxHeight: px,
-    size: px,
-    top: px,
-    right: px,
-    bottom: px,
-    left: px,
-
-    // Spacing props
-    padding: px,
-    paddingTop: px,
-    paddingRight: px,
-    paddingBottom: px,
-    paddingLeft: px,
-    margin: px,
-    marginTop: px,
-    marginRight: px,
-    marginBottom: px,
-    marginLeft: px,
-
-    // Transform props
-    rotate: degrees,
-    rotateX: degrees,
-    rotateY: degrees,
-    rotateZ: degrees,
-    scale,
-    scaleX: scale,
-    scaleY: scale,
-    scaleZ: scale,
-    skew: degrees,
-    skewX: degrees,
-    skewY: degrees,
-    distance: px,
-    translateX: px,
-    translateY: px,
-    translateZ: px,
-    x: px,
-    y: px,
-    z: px,
-    perspective: px,
-    transformPerspective: px,
-    opacity: alpha,
-    originX: progressPercentage,
-    originY: progressPercentage,
-    originZ: px,
-
-    // Misc
+    ...browserNumberValueTypes,
+    ...transformValueTypes,
     zIndex: int,
-    backgroundPositionX: px,
-    backgroundPositionY: px,
 
     // SVG
     fillOpacity: alpha,
