+++ packages/framer-motion/src/render/dom/value-types/transform.ts
@@ -1,0 +1,34 @@
+import { alpha, scale } from "../../../value/types/numbers"
+import {
+    degrees,
+    progressPercentage,
+    px,
+} from "../../../value/types/numbers/units"
+import { ValueTypeMap } from "./types"
+
+export const transformValueTypes: ValueTypeMap = {
+    rotate: degrees,
+    rotateX: degrees,
+    rotateY: degrees,
+    rotateZ: degrees,
+    scale,
+    scaleX: scale,
+    scaleY: scale,
+    scaleZ: scale,
+    skew: degrees,
+    skewX: degrees,
+    skewY: degrees,
+    distance: px,
+    translateX: px,
+    translateY: px,
+    translateZ: px,
+    x: px,
+    y: px,
+    z: px,
+    perspective: px,
+    transformPerspective: px,
+    opacity: alpha,
+    originX: progressPercentage,
+    originY: progressPercentage,
+    originZ: px,
+}
