+++ packages/framer-motion/src/value/types/__tests__/index.test.ts
@@ -1,4 +1,3 @@
-import { singleColorRegex, colorRegex, floatRegex } from "../utils"
 import { complex } from "../complex"
 import { hex } from "../color/hex"
 import { rgba, rgbUnit } from "../color/rgba"
@@ -7,6 +6,9 @@ import { color } from "../color"
 import { degrees, percent, progressPercentage, px } from "../numbers/units"
 import { alpha } from "../numbers"
 import { filter } from "../complex/filter"
+import { singleColorRegex } from "../utils/single-color-regex"
+import { colorRegex } from "../utils/color-regex"
+import { floatRegex } from "../utils/float-regex"
 
 const PATH = "M150 0 L75 200 L225 200 Z"
 const GREYSCALE = "greyscale(100%)"
