+++ packages/framer-motion/src/value/types/color/utils.ts
@@ -1,5 +1,7 @@
 import { Color, HSLA, RGBA } from "../types"
-import { floatRegex, isNullish, isString, singleColorRegex } from "../utils"
+import { floatRegex } from "../utils/float-regex"
+import { isNullish } from "../utils/is-nullish"
+import { singleColorRegex } from "../utils/single-color-regex"
 
 /**
  * Returns true if the provided string is a color, ie rgba(0,0,0,0) or #000,
@@ -7,7 +9,9 @@ import { floatRegex, isNullish, isString, singleColorRegex } from "../utils"
  */
 export const isColorString = (type: string, testProp?: string) => (v: any) => {
     return Boolean(
-        (isString(v) && singleColorRegex.test(v) && v.startsWith(type)) ||
+        (typeof v === "string" &&
+            singleColorRegex.test(v) &&
+            v.startsWith(type)) ||
             (testProp &&
                 !isNullish(v) &&
                 Object.prototype.hasOwnProperty.call(v, testProp))
@@ -17,7 +21,7 @@ export const isColorString = (type: string, testProp?: string) => (v: any) => {
 export const splitColor =
     <V extends RGBA | HSLA>(aName: string, bName: string, cName: string) =>
     (v: string | Color): V => {
-        if (!isString(v)) return v as any
+        if (typeof v !== "string") return v as any
 
         const [a, b, c, alpha] = v.match(floatRegex) as any
