+++ packages/framer-motion/src/value/types/utils/color-regex.ts
@@ -1,0 +1,2 @@
+export const colorRegex =
+    /(?:#[\da-f]{3,8}|(?:rgb|hsl)a?\((?:-?[\d.]+%?[,\s]+){2}-?[\d.]+%?\s*(?:[,/]\s*)?(?:\b\d+(?:\.\d+)?|\.\d+)?%?\))/giu
