+++ packages/framer-motion/src/value/types/utils/float-regex.ts
@@ -1,0 +1,1 @@
+export const floatRegex = /-?(?:\d+(?:\.\d+)?|\.\d+)/gu
