+++ packages/framer-motion/src/value/types/utils/is-nullish.ts
@@ -1,0 +1,3 @@
+export function isNullish(v: any): v is null | undefined {
+    return v == null
+}
